+++ poll-webapp/src/main/resources/locale/portlet/Poll_en.properties
@@ -2,5 +2,5 @@
 #                                    Create Poll Composer                           #
 #####################################################################################
 
-composer.poll.create = Create a Poll
-composer.poll.create.description = Add a Poll to your Activity
+composer.poll.create=Create a Poll
+composer.poll.create.description=Add a Poll to your Activity
