+++ poll-webapp/src/main/webapp/WEB-INF/gatein-resources.xml
@@ -5,9 +5,9 @@
   xmlns="http://www.exoplatform.org/xml/ns/gatein_resources_1_4">
  <portlet-skin>
     <application-name>poll</application-name>
-    <portlet-name>Poll</portlet-name>
+    <portlet-name>PollExtensions</portlet-name>
     <skin-name>Enterprise</skin-name>
-    <css-path>/skin/less/poll.less</css-path>
+    <css-path>/skin/css/poll.css</css-path>
   </portlet-skin>
 
   <portlet>
