+++ poll-webapp/src/main/webapp/WEB-INF/portlet.xml
@@ -2,11 +2,11 @@
 <portlet-app xmlns="http://java.sun.com/xml/ns/portlet/portlet-app_2_0.xsd"
              version="2.0"
              xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-             xsi:schemaLocation="http://java.sun.com/xml/ns/portlet/portlet-app_2_0.xsd http://java.sun.com/xml/ns/portlet/portlet-app_2_0.xsd">
+             xsi:schemaLocation="http://java.sun.com/xml/ns/portlet/portlet-app_2_0.xsd">
 
     <portlet>
       <portlet-name>PollExtensions</portlet-name>
-      <display-name xml:lang="EN">Poll Application Extensions</display-name>
+      <display-name>Poll Application Extensions</display-name>
       <portlet-class>org.exoplatform.commons.api.portlet.GenericDispatchedViewPortlet</portlet-class>
       <init-param>
       <name>portlet-view-dispatched-file-path</name>
