+++ poll-webapp/src/main/webapp/skin/less/poll.less
@@ -1,2 +1,17 @@
 @import "variables.less";
 @import "mixins.less";
+@images-path: "/poll/images";
+
+.composerActions {
+  .actionItem {
+    .actionItemIcon {
+      .createPollComposerIcon {
+        background: url('@{images-path}/create-poll-icon.png') no-repeat;
+        width: 60px;
+        height: 60px;
+        margin: 10px 0 0 10px;
+        background-size: 40px;
+      }
+    }
+  }
+}
