+++ poll-webapp/src/main/webapp/vue-app/poll-extensions/featureService.js
@@ -16,25 +16,15 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-import * as featureService from './featureService.js';
-
-let pollFeatureEnabled;
-
-featureService.isFeatureEnabled('poll').then(enabled => {
-  pollFeatureEnabled = enabled;
-});
-
-export function initExtensions() {
-  extensionRegistry.registerExtension('ActivityComposer', 'activity-composer-action', {
-    key: 'poll',
-    rank: 40,
-    resourceBundle: 'locale.portlet.Poll',
-    labelKey: 'composer.poll.create',
-    description: 'composer.poll.create.description',
-    iconClass: 'createPollComposerIcon',
-    enabled: pollFeatureEnabled,
-    onExecute: () => {
-      return null;
+export function isFeatureEnabled(featureName) {
+  return fetch(`${eXo.env.portal.context}/${eXo.env.portal.rest}/v1/features/${featureName}`, {
+    method: 'GET',
+    credentials: 'include',
+  }).then(resp => {
+    if (!resp || !resp.ok) {
+      throw new Error('Response code indicates a server error', resp);
+    } else {
+      return resp.text();
     }
-  });
+  }).then(featureEnabled => featureEnabled === 'true');
 }
