+++ poll-webapp/webpack.watch.js
@@ -3,9 +3,16 @@ const merge = require('webpack-merge');
 
 const webpackProductionConfig = require('./webpack.prod.js');
 
+// the display name of the war
+const app = 'poll';
+
+// add the server path to your server location path
+
+const exoServerPath = "/exo-server";
+
 module.exports = merge(webpackProductionConfig, {
   output: {
-    path: '/exo-server/webapps/poll/',
-    filename: 'js/[name].bundle.js'
-  }
+    path: path.resolve(`${exoServerPath}/webapps/${app}/`)
+  },
+  devtool: 'inline-source-map'
 });
